+++ crates/revmc-worker/src/lib.rs
@@ -1,8 +1,6 @@
-//! EVM runtime JIT worker implementation.
+//! EVM runtime AOT worker implementation.
 //!
-//! This crate provides the implementation for a JIT (Just-In-Time) compiler worker
-//! that operates in a node runtime environment. It includes modules for handling
-//! errors, external contexts, and worker functionalities.
+//! Jit
 
 #![allow(missing_docs)]
 pub mod error;
