+++ crates/revmc-worker/src/worker/mod.rs
@@ -3,7 +3,9 @@ mod compiler;
 mod hotcode;
 mod path;
 mod runtime;
+mod worker_pool;
 
-pub use compiler::*;
 pub(crate) use hotcode::*;
 pub(crate) use path::*;
+pub(crate) use runtime::*;
+pub use worker_pool::*;
